1.5-SNAPSHOT
- Added Repository (May add redis cache in the future)
- Added better console log formatting
- Added Hikari pool messages when throwing an exception
- Added statementing for Dao
- Change integer to long for identifiers
- Fixed bugs regarding Mysql (Haven't tested yet on mariadb, should work the same though)
- and a couple more bug fixes.

+++ src/main/java/com/huskydreaming/huskycore/interfaces/database/sql/SqlEntity.java
@@ -2,8 +2,8 @@ package com.huskydreaming.huskycore.interfaces.database.sql;
 
 public interface SqlEntity {
 
-    int getId();
-    void setId(int id);
+    long getId();
+    void setId(long id);
 
     SqlEntityType getEntityType();
 }
